// src/types/setting.type.ts
export type SettingType = {
  id?: number | null
  ofStatus?: string | null
  latitute?: number | null
  longtitute?: number | null
  distancePreference?: number | null
  lookFor?: string | null
  ageMin?: number | null
  ageMax?: number | null
  distanceUnit?: number | null
  globalMatches?: 0 | 1 | null
  hideAge?: 0 | 1 | null
  hideDistance?: 0 | 1 | null
  phoneNumber?: string | null
  email?: string | null
}